X86: If the caller clears the stack, don't let the callee modify the pushed paras.
Warning: Actually, this doubles stack-usage for paras in recursive functions, when it's declared as cdecl. The compiler actually copies the parameters even if they eren't modified in the body.

However, I personally do not want fpc to modify (on callee-side) parameters that are "caller-owned", since I frequently use cdecl callbacks.