This adds the collector thread to the Windows RTLs. As the new TLS variant with OS support is still in development and also won't work on all Windows variants (e.g. not on a vanilla Windows 98 SE) an alternate solution was needed.
Without further support by the OS only a kind of "garbage collector" was possible. When a thread that was not created by BeginThread of the application (!) accesses a threadvar for the first time (e.g. by accessing the standard IO) the RTL for that thread is initialized and the thread is registered with the collector.

The collector now periodically checks whether one of the registered threads has terminated (its handle is signaled) and "shuts down" the RTL for that thread by calling DoneThread while letting all threadvars point to the threadvar memory area of the terminated thread.

This isn't a beautiful solution, but I think that it is better than a memory leak.

Note for DLLs: the collector is not necessary here, because DLL_THREAD_ATTACH and DLL_THREAD_DETACH are sufficient. Also note that we might receive more DLL_THREAD_DETACH calls than DLL_THREAD_ATTACH ones, because threads that were running before the DLL was loaded don't call DLL_THREAD_ATTACH anymore.